+++ module-service/src/main/java/com/welcommu/moduleservice/project/dto/ProjectCreateRequest.java
@@ -26,7 +26,7 @@ public class ProjectCreateRequest {
     private String name;
     private String description;
 
-    private Long projectFee;
+    private int projectFee;
 
     @NotNull
     private LocalDate startDate;
